+++ Slim/Interfaces/MiddlewareCollectionInterface.php
@@ -5,10 +5,11 @@ declare(strict_types=1);
 namespace Slim\Interfaces;
 
 use Psr\Http\Server\MiddlewareInterface;
+use Slim\Routing\MiddlewareCollection;
 
 interface MiddlewareCollectionInterface
 {
-    public function getMiddlewareStack(): array;
+    // public function getMiddlewareStack(): MiddlewareCollection;
 
     public function add(MiddlewareInterface|callable|string $middleware): self;